refactor: wire JavaFxApp to shared createSyncModules factory, delete SyncClientComponents
