+++ platform-desktop-javafx/src/main/kotlin/io/github/rygel/outerstellar/platform/fx/app/JavaFxApp.kt
@@ -2,6 +2,7 @@ package io.github.rygel.outerstellar.platform.fx.app
 
 import io.github.rygel.outerstellar.i18n.I18nService
 import io.github.rygel.outerstellar.platform.analytics.NoOpAnalyticsService
+import io.github.rygel.outerstellar.platform.di.createSyncModules
 import io.github.rygel.outerstellar.platform.fx.FxAppContext
 import io.github.rygel.outerstellar.platform.fx.controller.MainController
 import io.github.rygel.outerstellar.platform.fx.service.FxStateProvider
@@ -31,13 +32,7 @@ import io.github.rygel.outerstellar.platform.sync.client.HttpSyncClient
 import io.github.rygel.outerstellar.platform.sync.client.NotificationClient
 import io.github.rygel.outerstellar.platform.sync.client.ProfileClient
 import io.github.rygel.outerstellar.platform.sync.client.SyncClient
-import io.github.rygel.outerstellar.platform.sync.engine.DefaultSessionLifecycle
 import io.github.rygel.outerstellar.platform.sync.engine.HttpConnectivityChecker
-import io.github.rygel.outerstellar.platform.sync.engine.module.AdminModuleImpl
-import io.github.rygel.outerstellar.platform.sync.engine.module.AuthModuleImpl
-import io.github.rygel.outerstellar.platform.sync.engine.module.NotificationModuleImpl
-import io.github.rygel.outerstellar.platform.sync.engine.module.ProfileModuleImpl
-import io.github.rygel.outerstellar.platform.sync.engine.module.SyncDataModuleImpl
 import io.micrometer.core.instrument.Metrics
 import java.awt.Desktop
 import java.net.URI
@@ -68,14 +63,6 @@ private class Repositories(
     val contactService: ContactService,
 )
 
-private class SyncModules(
-    val auth: AuthModuleImpl,
-    val syncData: SyncDataModuleImpl,
-    val profile: ProfileModuleImpl,
-    val admin: AdminModuleImpl,
-    val notification: NotificationModuleImpl,
-)
-
 class JavaFxApp : Application() {
 
     private lateinit var viewModel: FxSyncViewModel
@@ -85,7 +72,20 @@ class JavaFxApp : Application() {
         val clients = createHttpClients(fxConfig)
         val repositories = createRepositories(fxConfig)
         val analytics = NoOpAnalyticsService()
-        val modules = createModules(clients, repositories, analytics)
+        val modules =
+            createSyncModules(
+                syncClient = clients.sync,
+                authClient = clients.auth,
+                profileClient = clients.profile,
+                adminClient = clients.admin,
+                notificationClient = clients.notification,
+                messageService = repositories.messageService,
+                contactService = repositories.contactService,
+                analytics = analytics,
+                repository = repositories.messageRepository,
+                transactionManager = repositories.transactionManager,
+                notifier = FxTrayNotifier,
+            )
         val i18nService = I18nService.create("messages")
 
         viewModel =
@@ -140,43 +140,6 @@ class JavaFxApp : Application() {
         val messageService = MessageService(messageRepository, outboxRepository, transactionManager, NoOpMessageCache)
         val contactService = ContactService(contactRepository, NoOpEventPublisher, transactionManager)
         return Repositories(messageRepository, transactionManager, messageService, contactService)
-    }
-
-    private fun createModules(clients: HttpClients, repos: Repositories, analytics: NoOpAnalyticsService): SyncModules {
-        val lifecycle = DefaultSessionLifecycle()
-
-        val syncDataModule =
-            SyncDataModuleImpl(
-                syncClient = clients.sync,
-                messageService = repos.messageService,
-                contactService = repos.contactService,
-                analytics = analytics,
-                repository = repos.messageRepository,
-                transactionManager = repos.transactionManager,
-                lifecycle = lifecycle,
-                notifier = FxTrayNotifier,
-            )
-        val authModule =
-            AuthModuleImpl(
-                authClient = clients.auth,
-                analytics = analytics,
-                lifecycle = lifecycle,
-                notifier = FxTrayNotifier,
-            )
-        val profileModule =
-            ProfileModuleImpl(
-                profileClient = clients.profile,
-                analytics = analytics,
-                lifecycle = lifecycle,
-                notifier = FxTrayNotifier,
-            )
-        val adminModule = AdminModuleImpl(adminClient = clients.admin, analytics = analytics, lifecycle = lifecycle)
-        val notificationModule =
-            NotificationModuleImpl(notificationClient = clients.notification, lifecycle = lifecycle)
-
-        lifecycle.initialize(syncDataModule = syncDataModule, authModule = authModule, authClient = clients.auth)
-
-        return SyncModules(authModule, syncDataModule, profileModule, adminModule, notificationModule)
     }
 
     override fun start(primaryStage: Stage) {
