+++ platform-sync-client/src/main/kotlin/io/github/rygel/outerstellar/platform/di/SyncModules.kt
@@ -6,13 +6,7 @@ import io.github.rygel.outerstellar.platform.persistence.TransactionManager
 import io.github.rygel.outerstellar.platform.service.ContactService
 import io.github.rygel.outerstellar.platform.service.MessageService
 import io.github.rygel.outerstellar.platform.sync.client.AdminClient
-import io.github.rygel.outerstellar.platform.sync.client.ApiSession
 import io.github.rygel.outerstellar.platform.sync.client.AuthClient
-import io.github.rygel.outerstellar.platform.sync.client.HttpAdminClient
-import io.github.rygel.outerstellar.platform.sync.client.HttpAuthClient
-import io.github.rygel.outerstellar.platform.sync.client.HttpNotificationClient
-import io.github.rygel.outerstellar.platform.sync.client.HttpProfileClient
-import io.github.rygel.outerstellar.platform.sync.client.HttpSyncClient
 import io.github.rygel.outerstellar.platform.sync.client.NotificationClient
 import io.github.rygel.outerstellar.platform.sync.client.ProfileClient
 import io.github.rygel.outerstellar.platform.sync.client.SyncClient
@@ -29,42 +23,29 @@ import io.github.rygel.outerstellar.platform.sync.engine.module.ProfileModule
 import io.github.rygel.outerstellar.platform.sync.engine.module.ProfileModuleImpl
 import io.github.rygel.outerstellar.platform.sync.engine.module.SyncDataModule
 import io.github.rygel.outerstellar.platform.sync.engine.module.SyncDataModuleImpl
-import org.http4k.client.JavaHttpClient
-import org.http4k.core.HttpHandler
 
-class SyncClientComponents(
-    val session: ApiSession,
-    val authClient: AuthClient,
-    val syncClient: SyncClient,
-    val profileClient: ProfileClient,
-    val adminClient: AdminClient,
-    val notificationClient: NotificationClient,
-    val authModule: AuthModule,
-    val syncDataModule: SyncDataModule,
-    val profileModule: ProfileModule,
-    val adminModule: AdminModule,
-    val notificationModule: NotificationModule,
+class SyncModules(
+    val auth: AuthModule,
+    val syncData: SyncDataModule,
+    val profile: ProfileModule,
+    val admin: AdminModule,
+    val notification: NotificationModule,
 )
 
-fun createSyncClientComponents(
-    baseUrl: String,
-    analytics: AnalyticsService,
+fun createSyncModules(
+    syncClient: SyncClient,
+    authClient: AuthClient,
+    profileClient: ProfileClient,
+    adminClient: AdminClient,
+    notificationClient: NotificationClient,
     messageService: MessageService,
     contactService: ContactService?,
+    analytics: AnalyticsService,
     repository: MessageRepository,
     transactionManager: TransactionManager,
     notifier: ModuleNotifier? = null,
     connectivityChecker: ConnectivityChecker? = null,
-): SyncClientComponents {
-    val session = ApiSession()
-    val httpClient: HttpHandler = JavaHttpClient()
-
-    val authClient = HttpAuthClient(baseUrl, session, httpClient)
-    val syncClient = HttpSyncClient(baseUrl, session, httpClient)
-    val profileClient = HttpProfileClient(baseUrl, session, httpClient)
-    val adminClient = HttpAdminClient(baseUrl, session, httpClient)
-    val notificationClient = HttpNotificationClient(baseUrl, session, httpClient)
-
+): SyncModules {
     val lifecycle = DefaultSessionLifecycle()
 
     val syncDataModule: SyncDataModule =
@@ -79,10 +60,8 @@ fun createSyncClientComponents(
             notifier = notifier,
             connectivityChecker = connectivityChecker,
         )
-
     val authModule: AuthModule =
         AuthModuleImpl(authClient = authClient, analytics = analytics, lifecycle = lifecycle, notifier = notifier)
-
     val profileModule: ProfileModule =
         ProfileModuleImpl(
             profileClient = profileClient,
@@ -90,26 +69,12 @@ fun createSyncClientComponents(
             lifecycle = lifecycle,
             notifier = notifier,
         )
-
     val adminModule: AdminModule =
         AdminModuleImpl(adminClient = adminClient, analytics = analytics, lifecycle = lifecycle)
-
     val notificationModule: NotificationModule =
         NotificationModuleImpl(notificationClient = notificationClient, lifecycle = lifecycle)
 
     lifecycle.initialize(syncDataModule = syncDataModule, authModule = authModule, authClient = authClient)
 
-    return SyncClientComponents(
-        session = session,
-        authClient = authClient,
-        syncClient = syncClient,
-        profileClient = profileClient,
-        adminClient = adminClient,
-        notificationClient = notificationClient,
-        authModule = authModule,
-        syncDataModule = syncDataModule,
-        profileModule = profileModule,
-        adminModule = adminModule,
-        notificationModule = notificationModule,
-    )
+    return SyncModules(authModule, syncDataModule, profileModule, adminModule, notificationModule)
 }
